test: split down test stuff a bit
